+++ android/src/main/java/io/ltbl/bdkrn/Utils.kt
@@ -60,9 +60,11 @@ fun setAddressIndex(addressIndex: Any?): AddressIndex {
             "lastUnused" -> AddressIndex.LastUnused
             else -> AddressIndex.New
         }
+
         is Double -> {
             AddressIndex.Peek(addressIndex.toUInt())
         }
+
         else -> AddressIndex.New
     }
 }
@@ -71,13 +73,13 @@ fun randomId() = UUID.randomUUID().toString()
 
 fun getTransactionObject(transaction: TransactionDetails): MutableMap<String, Any> {
     return mutableMapOf<String, Any>(
-        "received" to transaction.received.toInt(),
-        "sent" to transaction.sent.toInt(),
-        "fee" to transaction.fee!!.toInt(),
+        "received" to transaction.received.toDouble(),
+        "sent" to transaction.sent.toDouble(),
+        "fee" to transaction.fee!!.toDouble(),
         "txid" to transaction.txid,
         "confirmationTime" to mutableMapOf<String, Any>(
             "height" to (transaction.confirmationTime?.height?.toInt() ?: 0),
-            "timestamp" to (transaction.confirmationTime?.timestamp?.toInt() ?: 0),
+            "timestamp" to (transaction.confirmationTime?.timestamp?.toDouble() ?: 0),
         )
     )
 }
@@ -144,14 +146,14 @@ fun getPayload(payload: Payload): MutableMap<String, Any> {
 fun createTxOut(txOut: TxOut, _scripts: MutableMap<String, Script>): MutableMap<String, Any> {
     val randomId = randomId()
     _scripts[randomId] = txOut.scriptPubkey
-    return mutableMapOf("script" to randomId, "value" to txOut.value.toInt())
+    return mutableMapOf("script" to randomId, "value" to txOut.value.toDouble())
 }
 
 fun createTxIn(txIn: TxIn, _scripts: MutableMap<String, Script>): MutableMap<String, Any> {
     val randomId = randomId()
     _scripts[randomId] = txIn.scriptSig
     var witnessList = mutableListOf<Any>();
-    for(item in txIn.witness) {
+    for (item in txIn.witness) {
         witnessList.add(makeNativeArray(item))
     }
     return mutableMapOf(