+++ lib/classes/NativeLoader.d.ts
@@ -26,7 +26,7 @@ export interface NativeBdkRn {
     sledDBInit(path: string, treeName: string): string;
     sqliteDBInit(path: string): string;
     walletInit(descriptor: string, changeDescriptor: string | null, network: Network, dbConfig: string): any;
-    getAddress(id: string, addressIndex: AddressIndex): any;
+    getAddress(id: string, addressIndex: AddressIndex | number): any;
     getInternalAddress(id: string, addressIndex: AddressIndex | number): any;
     isMine(id: string, scriptId: string): boolean;
     getBalance(id: string): Balance;
@@ -35,13 +35,14 @@ export interface NativeBdkRn {
     sign(id: string, psbtBase64: string, signOptions?: SignOptions): string;
     listUnspent(id: string): Array<any>;
     listTransactions(id: string, includeRaw: boolean): Array<TransactionDetails>;
-    initAddress(address: string): string;
+    initAddress(address: string, network: string): string;
     addressFromScript(script: string, network: Network): string;
     addressToScriptPubkeyHex(id: string): string;
     addressPayload(id: string): Payload;
     addressNetwork(id: string): string;
     addressToQrUri(id: string): string;
     addressAsString(id: string): string;
+    addressIsValidForNetwork(id: string, network: string): boolean;
     createTxBuilder(): string;
     addRecipient(id: string, scriptId: string, amount: number): string;
     finish(id: string, walletId: string): {
@@ -69,9 +70,11 @@ export interface NativeBdkRn {
     newBip44(id: string, keychain: KeychainKind, network: Network): string;
     newBip49(id: string, keychain: KeychainKind, network: Network): string;
     newBip84(id: string, keychain: KeychainKind, network: Network): string;
+    newBip86(id: string, keychain: KeychainKind, network: Network): string;
     newBip44Public(id: string, fingerprint: string, keychain: KeychainKind, network: Network): string;
     newBip49Public(id: string, fingerprint: string, keychain: KeychainKind, network: Network): string;
     newBip84Public(id: string, fingerprint: string, keychain: KeychainKind, network: Network): string;
+    newBip86Public(id: string, fingerprint: string, keychain: KeychainKind, network: Network): string;
     combine(psbt64: string, otherPsbt: string): string;
     extractTx(psbt64: string): string;
     serialize(psbt64: string): string;
@@ -80,7 +83,7 @@ export interface NativeBdkRn {
     psbtFeeRate(psbt64: string): number;
     jsonSerialize(psbt64: string): string;
     bumpFeeTxBuilderInit(txid: string, newFeeRate: number): string;
-    bumpFeeTxBuilderAllowShrinking(id: string, address: string): string;
+    bumpFeeTxBuilderAllowShrinking(id: string, scriptId: string): string;
     bumpFeeTxBuilderEnableRbf(id: string): any;
     bumpFeeTxBuilderEnableRbfWithSequence(id: string, nsequence: number): any;
     bumpFeeTxBuilderFinish(id: string, walletId: string): any;